requirements.txt
Flask==2.3.3
pycryptodome==3.20.0
httpx==0.27.0
protobuf==4.25.3